+++ xcode/Subconscious/Shared/Components/Deck/CardStackView.swift
@@ -1,132 +1,5 @@
 import SwiftUI
 
-struct CardContentView: View {
-    @Environment (\.colorScheme) var colorScheme
-    var entry: EntryStub
-    var backlinks: [EntryStub]
-    var onLink: (EntryLink) -> Void
-    
-    var highlight: Color {
-        entry.highlightColor(colorScheme: colorScheme)
-    }
-    
-    
-    var body: some View {
-        SubtextView(
-            peer: entry.toPeer(),
-            subtext: entry.excerpt,
-            onLink: onLink
-        )
-        // Opacity allows blendMode to show through
-        .foregroundStyle(.primary.opacity(0.8))
-        .accentColor(highlight)
-        .padding(DeckTheme.cardPadding)
-        
-        Spacer()
-        
-        HStack {
-            Text(
-                entry.address.markup
-            )
-            .lineLimit(1)
-            
-            if !backlinks.isEmpty {
-                Spacer()
-                
-                HStack {
-                    Image(systemName: "link")
-                    Text("\(backlinks.count)")
-                }
-            }
-        }
-        .font(.caption)
-        .foregroundStyle(highlight)
-        .padding(DeckTheme.cardPadding)
-    }
-}
-
-struct CardView: View {
-    @Environment(\.colorScheme) private var colorScheme
-    
-    var entry: CardModel
-    var onLink: (EntryLink) -> Void
-    
-    var color: Color {
-        switch entry.card {
-        case let .entry(entry, _, _):
-            return entry.color(colorScheme: colorScheme)
-        case let .prompt(_, entry, _, _):
-            return entry.color(colorScheme: colorScheme)
-        default:
-            return .secondary
-        }
-    }
-    
-    var highlight: Color {
-        switch entry.card {
-        case let .entry(entry, _, _):
-            return entry.highlightColor(colorScheme: colorScheme)
-        case let .prompt(_, entry, _, _):
-            return entry.highlightColor(colorScheme: colorScheme)
-        default:
-            return .secondary
-        }
-    }
-    
-    var blendMode: BlendMode {
-        colorScheme == .dark
-           ? .plusLighter
-           : .plusDarker
-    }
-    
-    var body: some View {
-        VStack(alignment: .leading, spacing: AppTheme.unit2) {
-            switch entry.card {
-            case let .entry(entry, _, backlinks):
-                CardContentView(
-                    entry: entry,
-                    backlinks: backlinks,
-                    onLink: onLink
-                )
-            case let .prompt(message, entry, _, backlinks):
-                VStack(alignment: .leading, spacing: 0) {
-                    HStack {
-                        Image(systemName: "sparkles")
-                        
-                        Text(message)
-                    }
-                    .padding(DeckTheme.cardPadding)
-                    .foregroundStyle(highlight)
-                    .font(.subheadline)
-                    
-                    Divider()
-                        .overlay(highlight)
-                    
-                    CardContentView(
-                        entry: entry,
-                        backlinks: backlinks,
-                        onLink: onLink
-                    )
-                }
-            case .action(let string):
-                // TEMP
-                VStack {
-                    Image(systemName: "scribble.variable")
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .frame(width: 64, height: 64)
-                    Text(string)
-                }
-                .foregroundStyle(.secondary)
-            }
-            
-        }
-        .allowsHitTesting(false)
-        .background(color)
-        .cornerRadius(DeckTheme.cornerRadius)
-    }
-}
-
 enum CardDragGestureProgress {
     case inactive
     case started
@@ -190,7 +63,7 @@ struct CardEffectModifier: ViewModifier {
             // Reduce shadow intensity with depth
             .shadow(
                 color: DeckTheme.cardShadow.opacity(
-                    0.2 * (1.0 - 5.0*stackFactor)
+                    0.15 * (1.0 - 5.0*stackFactor)
                 ),
                 radius: 2.5,
                 x: 0,
